Fix sequential writes for streams (tcp, output).
Strands do not make async_write work reasonably (that is finish one full write before starting another). It would be really nice if Boost.Asio could provide this very basic abstraction, instead of all the other obnoxious ones.
Now output and tcp sockets use an internal buffer where data is appended and consumed sequentially. It is no longer zero-copy, but it can actually be a speedup since multiple small writes can be coalesced in a larger write to the output.